Rename logo and add 32x32 version
