+++ src/js/components/Bubble.js
@@ -1,6 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
-import logoUrl from "images/logo.png"
+import logoUrl from "images/moco-159x159.png"
 
 const Bubble = ({ bookedHours }) => (
   <div className="moco-bx-bubble-inner">
